README.md
Before we start, please make sure that your environment is correctly set up and import all the dependencies.
Now, we will load and preprocess our data. We will use the federated version of MNIST data.
We preprocessed the data and sample the set of clients once, and reuse the same set across rounds to speed up convergence. In the next step, we will create our model.
Now it’s time to train our model on decentralized or user data.
This step will complete your model training on client or user data. Finally, we will see how we can evaluate and test the model.
So, this is how you can train your centralized machine learning model on decentralized or user-sensitive data!